feat: Complete colocation restructure - move page components to their respective directories
- Move all page-specific components to app/[page]/components/
- Migrate about, articles, contact, works components to colocation structure
- Move top/service-list to features/top/components/
- Update all import statements to use relative imports within pages
- Update cross-page imports to use new paths
- Remove empty components/ directory completely
- Build passing ✓

Colocation structure:
- app/about/components/about-sns-box.tsx
- app/articles/components/articles-body.tsx, articles-list.tsx
- app/contact/components/contact-body.tsx, input-text.tsx
- app/works/components/works-list.tsx
- app/works/[slug]/components/works-detail-*.tsx
- features/top/components/service-list.tsx

This completes the Colocation + Feature-based hybrid architecture migration.
