+++ src/app/about/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next'
 import { ImageWrapper } from '@/features/ui/image-wrapper'
 import { TextBtn } from '@/features/ui/text-btn'
 import { Title } from '@/features/ui/title'
-import { AboutSnsBox } from '@/components/pages/about/about-sns-box'
+import { AboutSnsBox } from './components/about-sns-box'
 
 import { setMetaData } from '@/utils'
