+++ src/app/works/[slug]/components/works-detail-body.tsx
