+++ src/app/works/[slug]/components/works-detail-info.tsx
