+++ src/app/works/[slug]/components/works-slider.tsx
